Refactor course API

+++ src/main/java/com/team/updevic001/model/dtos/response/course/ResponseCategoryDto.java
@@ -12,6 +12,6 @@ public class ResponseCategoryDto {
 
     private CourseCategoryType category;
 
-    private Integer courseCount;
+    private long courseCount;
 
 }
